Cleaning up entities reducer code (#587)
